Complete foundations/flex/07-flex-layout-2
At first I was using align-content to center the header and footer, but
then I realized that it shouldn't be working beccause that property is
specifically for multi-line wrap elements. I did some reading and saw
that it was bad practice to use that, even though the browser allowed
it. So, I changed those two elemment's display properties to flex and
used align-items instead.